% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_etc.R
\name{as_numeric_without_warnings}
\alias{as_numeric_without_warnings}
\title{Wrapper function to convert into numeric without warnings}
\usage{
as_numeric_without_warnings(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A numeric or NA.
}
\description{
Wrapper function to convert into numeric without warnings
}
