% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{clean_station}
\alias{clean_station}
\title{Clean up station information}
\usage{
clean_station(station)
}
\arguments{
\item{station}{A String of station information.}
}
\value{
A tibble including station information.
}
\description{
Helper function for download_climate().
}
\examples{
data(station_links)
station_links \%>\%
  head(1) \%>\%
  `$`("station") \%>\%
  stringi::stri_unescape_unicode() \%>\%
  clean_station()

}
