% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{download_climate}
\alias{download_climate}
\alias{donwload_climate}
\title{Donwload climate data of the world}
\usage{
donwload_climate(url)
}
\arguments{
\item{url}{A String to specify target html.}
}
\value{
A tibble including climate and station information.
}
\description{
For polite scraping, 5 sec interval is set in donwload_climate(),
it takes over 5 hours to get climate data of all stations.
Please use existing links by "data(world_climate)",
if you do not need to renew climate data.
You can see web page as below.
https://www.data.jma.go.jp/gmd/cpd/monitor/nrmlist/
}
\examples{
# If you want all climate data, remove head().
library(tidyverse)
data(station_links)
station_links <-
  station_links \%>\%
  head() \%>\%
  `$`("url")

\dontrun{
climate <- list()
for(i in seq_along(station_links)){
  print(stringr::str_c(i, " / ", length(station_links)))
  climate[[i]] <- donwload_climate(station_links[i])
}
world_climate <- dplyr::bind_rows(climate)
world_climate
}

}
