% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gracefully_fail}
\alias{gracefully_fail}
\title{Graceful fail}
\usage{
gracefully_fail(remote_file)
}
\arguments{
\item{remote_file}{A string of remote file.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ An XML document when successed, or invisible NULL when failed.
}\if{html}{\out{</div>}}
}
\description{
Graceful fail
}
\references{
\url{https://gist.github.com/kvasilopoulos/47f24348ed75cdb6365312b17f4b914c}
}
