% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_etc.R
\name{head_3}
\alias{head_3}
\title{Wrapper function to head 3 items}
\usage{
head_3(x)
}
\arguments{
\item{x}{An object.}
}
\value{
An object like x with length 3.
}
\description{
Wrapper function to head 3 items
}
