% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_jp_full}
\alias{climate_jp_full}
\alias{climate_jp_full_tmp}
\title{Climate normals of all stations in Japan}
\format{
A data frame with 21697 rows (1669 stations * 13 rows)
and 21 variables. All variables are character.
\describe{
\item{month                 }{Month (1 to 12) or yearly value. Escaped by stringi::stri_escape_unicode().}
\item{air_pressure_land     }{Mean air pressure on land (hPa)}
\item{air_pressure_sea      }{Mean air pressure on sea level (hPa)}
\item{precipitation         }{Mean precipitation (mm)}
\item{temperature           }{Mean temperature (degree Celsius)}
\item{temperature_max       }{Mean of daily maximum temperature (degree Celsius)}
\item{temperature_min       }{Mean of daily minimum temperature (degree Celsius)}
\item{steam_pressure        }{Mean steam pressure (hPa)}
\item{moisture              }{Mean relative humidity (percent)}
\item{wind_speed            }{Mean wind speed (m/s)}
\item{wind_direction        }{Most frequent wind direction. Escaped by stringi::stri_escape_unicode().}
\item{sunshine              }{Mean sunshine duration (hour)}
\item{global_solar_radiation}{Mean global solar radiation (MJ/m^2)}
\item{snow_fall             }{Mean snow fall (cm)}
\item{snow_fall_max_per_day }{Mean of daily maximum snow fall (cm)}
\item{snow_deepest          }{Mean of deepest snow (cm)}
\item{cloud_cover           }{Mean cloud cover}
\item{days_of_snow          }{Mean days of snow}
\item{days_of_fog           }{Mean days of fog}
\item{days_of_thunderstorm  }{Mean days of thunderstorm}
\item{station_no            }{Station no. Can be joined to station_no of data(station_jp_full). Escaped by stringi::stri_escape_unicode().}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21697 rows and 21 columns.
}
\usage{
climate_jp_full

climate_jp_full_tmp
}
\description{
Monthly normals downloaded from Japan Meteorological Agency web pages
for all stations in data(station_jp_full).
Values are raw strings of the web pages,
because they include marks of 'JMA'
("@", ")", "]" and "#" show quality of the value,
"///" and "---" show no value).
Use as_numeric_without_warnings() after removing the marks,
or use download_detail(), which cleans the marks.
Stations of "AMeDAS" have no value in 11 of the columns,
because only 10 items are observed.
https://www.data.jma.go.jp/obd/stats/etrn/index.php
}
\examples{
library(magrittr)
library(stringi)
library(dplyr)
data(climate_jp_full)
climate_jp_full \%>\%
  head_3() \%>\%
  dplyr::mutate_if(is.character, stringi::stri_unescape_unicode)
}
\keyword{datasets}
