% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_detail.R
\name{detail_url}
\alias{detail_url}
\alias{detail_item}
\title{Build urls of detail climate data in Japan}
\usage{
detail_url(
  prec_no,
  block_no,
  item = "daily",
  year = NULL,
  month = NULL,
  day = NULL
)

detail_item(item, block_no)
}
\arguments{
\item{prec_no}{A string or numeric of area no.}

\item{block_no}{A string or numeric of station no.
Keep leading zero of 'AMeDAS' as a string (ex. "0588").}

\item{item}{A string to specify data type. See details.}

\item{year}{A numeric or string of year.}

\item{month}{A numeric or string of month.}

\item{day}{A numeric or string of day.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A string vector of urls.
}\if{html}{\out{</div>}}
}
\description{
Japan Meteorological Agency ('JMA') provides detail climate data
(yearly, monthly, daily, hourly and 10 minutes values) for each station.
detail_url() builds urls of the data,
and download_detail() downloads the data of the url.
You can see web page as below.
https://www.data.jma.go.jp/obd/stats/etrn/index.php
}
\details{
A station is specified by a pair of prec_no (area) and block_no (station).
Use download_prec_no() and download_block_no() to get the numbers.
Note that block_no of "kansho" (weather station) has 5 digits (ex. "47759")
and that of "AMeDAS" has 4 digits (ex. "0588").
Suffix of item ("_s" or "_a") is set to match the type of block_no.

item can be a short name in the following list,
or a name of php file without extension (ex. "daily_s1").
\describe{
\item{annually   }{Yearly values.}
\item{monthly    }{Monthly values in a year (year is required).}
\item{monthly_all}{Monthly values of all years ("kansho" only).}
\item{daily      }{Daily values in a month (year and month are required).}
\item{hourly     }{Hourly values in a day (year, month and day are required).}
\item{10min      }{10 minutes values in a day (year, month and day are required).}
\item{nml_ym     }{Normals of each month.}
\item{nml_daily  }{Normals of each day in a month (month is required).}
\item{nml_season }{Normals of first and last day of frost, snow and ice ("kansho" only).}
}
}
\examples{
detail_url(61, 47759, "daily", 2023, 6)
detail_url(61, "0588", "hourly", 2023, 6, 20)
detail_url(61, 47759, "daily", 2023, 1:3)

}
