% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_detail.R
\name{download_detail}
\alias{download_detail}
\title{Download detail climate data in Japan}
\usage{
download_detail(url, as_numeric = TRUE)
}
\arguments{
\item{url}{A string to specify target html.}

\item{as_numeric}{A logical. If TRUE, columns of numbers are
converted into numeric. Marks of 'JMA'
for missing values are treated as NA.}
}
\value{
A tibble including detail climate data, or NULL when failed.
}
\description{
For polite scraping, 5 sec interval is set in download_detail().
Please do not download too many data at once.
Use detail_url() to build an url of the data.
}
\examples{
\donttest{
url <- detail_url(61, 47759, "daily", 2023, 6)
download_detail(url)
}
}
