% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_detail.R
\name{download_prec_no}
\alias{download_prec_no}
\alias{download_block_no}
\title{Download prec_no and block_no of stations in Japan}
\usage{
download_prec_no()

download_block_no(prec_no)
}
\arguments{
\item{prec_no}{A string or numeric of area no.}
}
\value{
download_prec_no(): a tibble of prec_no and area.
download_block_no(): a tibble of prec_no, block_no, station and type.
NULL when failed.
}
\description{
A station of detail climate data is specified by a pair of
prec_no (area) and block_no (station).
download_prec_no() downloads all prec_no,
and download_block_no() downloads block_no of the area.
Please use existing data by "data(station_jp_full)",
which includes prec_no and station_no of all stations,
if you do not need to renew them.
For polite scraping, 5 sec interval is set in both functions.
You can see web page as below.
https://www.data.jma.go.jp/obd/stats/etrn/index.php
}
\examples{
\donttest{
prec_no <- download_prec_no()
prec_no
download_block_no(61)
}
}
