% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean_cli}
\alias{mean_cli}
\title{Downloaded tables of climate normals in Japan}
\format{
A list of 1673 tibbles.
\describe{
\item{21 columns}{157 tables of "kansho" (weather station).}
\item{10 columns}{1512 tables of "AMeDAS".}
\item{NULL      }{4 stations, which have no normals.}
}
}
\usage{
mean_cli
}
\description{
Raw tables downloaded from Japan Meteorological Agency web pages,
before cleaning into data(climate_jp_full).
Kept for reproducibility of data-raw/climate_jp_full.R.
Each element is a table of a station of data(station_jp_full),
and the order of the elements is the same as the rows of the data.
Column names are not set, and the first rows are headers of the web page.
https://www.data.jma.go.jp/obd/stats/etrn/index.php
}
\examples{
data(mean_cli)
length(mean_cli)
dim(mean_cli[[1]])
}
\keyword{datasets}
