% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_jp}
\alias{station_jp}
\title{Climate stations in Japan}
\format{
A data frame with 157 rows and 11 variable:
\describe{
\item{region       }{Region in Japan}
\item{pref         }{Prefecture in Japan}
\item{no           }{Station no}
\item{station      }{Station name. To avoid duplication, including country name after station name. Can split by "_". Escaped by stringi::stri_escape_unicode().}
\item{altitude     }{Altitude (m)}
\item{latitude     }{Latitude (degree)}
\item{NS           }{North or South.}
\item{longitude    }{Longitude (degree)}
\item{WE           }{West or East}
\item{yomi         }{Japanese pronunciation of station}
\item{city         }{Location of station}
}
}
\usage{
station_jp
}
\description{
Climate stations in Japan
}
\examples{
library(tidyverse)
data(station_jp)
station_jp \%>\%
  dplyr::mutate_all(stringi::stri_unescape_unicode)
}
\keyword{datasets}
