% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_jp_full}
\alias{station_jp_full}
\title{All stations in Japan}
\format{
A data frame with 1673 rows and 16 variables:
\describe{
\item{station      }{Station name. Escaped by stringi::stri_escape_unicode().}
\item{prec_no      }{Area no of 'JMA'. Escaped by stringi::stri_escape_unicode().}
\item{station_no   }{Station no of 'JMA', which is used as block_no in detail_url(). Escaped by stringi::stri_escape_unicode().}
\item{yomi         }{Pronunciation in Japanese. Escaped by stringi::stri_escape_unicode().}
\item{altitude     }{Altitude (m)}
\item{precipitation}{Flag of observation (1: observed, 0: not observed). Character, not numeric.}
\item{wind         }{Flag of observation (1: observed, 0: not observed)}
\item{temperature  }{Flag of observation (1: observed, 0: not observed)}
\item{sunshine     }{Flag of observation (1: observed, 0: not observed)}
\item{snow         }{Flag of observation (1: observed, 0: not observed)}
\item{moisture     }{Flag of observation (1: observed, 0: not observed)}
\item{prec         }{Area name of prec_no. Escaped by stringi::stri_escape_unicode().}
\item{latitude     }{Latitude (degree)}
\item{longitude    }{Longitude (degree)}
\item{end_date     }{Date of the end of observation. "9999-99-99" shows that the station is in operation (1287 stations). Escaped by stringi::stri_escape_unicode().}
\item{memo         }{Note of 'JMA'. Escaped by stringi::stri_escape_unicode().}
}
}
\usage{
station_jp_full
}
\description{
Station information downloaded from Japan Meteorological Agency web pages.
data(station_jp) includes only 157 stations of climate data,
while this data includes all 1673 stations.
A pair of prec_no and station_no specifies a station
in detail_url() and download_detail().
Note that station_no of "kansho" (weather station) has 5 digits (159 stations)
and that of "AMeDAS" has 4 digits (1514 stations).
https://www.data.jma.go.jp/obd/stats/etrn/index.php
}
\examples{
library(magrittr)
library(stringi)
library(dplyr)
data(station_jp_full)
station_jp_full \%>\%
  head_3() \%>\%
  dplyr::mutate_if(is.character, stringi::stri_unescape_unicode)
}
\keyword{datasets}
