% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_links}
\alias{station_links}
\title{Station name and its URL}
\format{
A data frame with  3444 rows and 3 variable:
\describe{
\item{no           }{Id number}
\item{station      }{Station information including no, month, temperature, precipitation, station, country, latitude, NS, longitude, WE, altitude. The information is NOT cleaned Row information donwloaded from each URL. Escaped by stringi::stri_escape_unicode().}
\item{link         }{URL of station.}
}
}
\usage{
station_links
}
\description{
Station name and its URL
}
\examples{
library(tidyverse)
data(station_links)
station_links \%>\%
  dplyr::mutate_all(stringi::stri_unescape_unicode)
}
\keyword{datasets}
